fix(code-rabbit): add ISO-3166 alpha-2 validation and int constraints
Applies 3 of 4 CodeRabbit findings (rejected #1 — re-adding .strict()
to signals.ts would undo the previous Zod v4 brand fix):

- Extract isoAlpha2 regex constant and apply to both alpha2 sites
- Use z.number().int().min(0) for memberCount (3 sites)
- Use z.number().int().min(0) for contributingSignals (1 site)

+++ packages/domain/src/risk/scoring.ts
@@ -23,7 +23,7 @@ export const riskScoreSchema = z
     score: z.number().min(0).max(100),
     level: riskLevelSchema,
     worstSeverity: signalSeveritySchema.nullable(),
-    contributingSignals: z.number().min(0),
+    contributingSignals: z.number().int().min(0),
   })
   .strict();
